requirements.txt

Flask==2.0.1
flask-cors==3.0.10
psutil==5.9.0
pyOpenSSL==23.2.0
